feat:Websockt fctn to notify user when msg is seen
Web sockt servers listens to "seenMessage" event and calls the function emitSeenMsgToUsers
The emitSeenMsgToUsers function sends a "seenMessage" event to a list of socket IDs, indicating that a message in a conversation has been seen. The function iterates through the provided socket IDs, emitting the event along with the message and conversation details. Additionally, it logs a message to the console for each notification sent.