change: add members (socket part)
When a user A added a user B to a conversation.
1)It first sent the request to the backend.
2)if the operation with db succeed, it returned to user A the updated conv state in db.
3)user A sent it back to backend through sockets,
4)Backend  emitted it to the concerned sockets

I learned that,instead of doing this way. It would be better that, at step 2, instead of returning it to user A, send it instantly to user Sockets as i did there : https://github.com/SebastienBotty/messenger-clone--Backend/commit/2a9834bad642e553a6c2b6940dc6a539154c737c

But somehow, when i do it this way, user A gets an error(Why? He receives the same event and data as other users)
So i decided to res.json(data) to user A AND to instant socket emit to other conv user than him.